refacto(excel): cleanup
